docs: README video demo (#16)
## 📜 Description

Added video in readme to show what package is doing.

## 💡 Motivation and Context

Many developers are not familiar with the Portal or teleport concept, so
they need to read a lot of text before they get an understanding what
this package is dong etc.

Sometimes it's easier to watch a video once and get the idea, so I've
tried to make a showcase of
https://github.com/kirillzyusko/react-native-teleport/pull/13 The idea
is that we "teleport" an animated sticker through layers and make it
non-blurrable.

I'm not sure it's the great idea to show this, because people may
associuate this video with blur or with other libs, but 3D dimension
video should show the idea behind the concept 🤞

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### Docs

- added a video demonstrating the possibility of the library;

## 🤔 How Has This Been Tested?

Tested manually via preview.

## 📸 Screenshots (if appropriate):

<!-- Add screenshots/video if needed -->
<!-- That would be highly appreciated if you can add how it looked
before and after your changes -->

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed