feat: `usePortal` hook (#17)
## 📜 Description

Added `usePortal` hook with ability to remove portal imperatively.

## 💡 Motivation and Context

Sometimes it can be useful to delete all mounted portals - for example
you show a popover and by screen click you should close it. In a
canonical way you need to add state manager and don't render portal if
state manager prevents it. But I decided to pack this functionality in
this package.

I've been thinking a lot how would be better to add it - and decided to
stick to this schema:
- imperative call takes control over declarative one;
- we'll hold reference to each `Portal` in js world (using `useId`
hook);
- we have 1:M relationship (one `PortalHost` can hold several `Portal`).

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- added `PortalManagerContext`;
- cover new functionality with unit-tests;
- added new `useId` hook for internal usage;
- added state manager based on `useReducer` for managing `Portal`
visibility;

## 🤔 How Has This Been Tested?

Tested in example app.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed