feat: flex styles example (#26)
## 📜 Description

Added an example when teleported child has `flex: 1` style.

## 💡 Motivation and Context

The original idea is that of having this example is that when we
teleport a view it should use parent layout and don't be restricted of
its original location. This indeed works on web, but doesn't work on
mobile yet. There is an ongoing investigation
https://github.com/kirillzyusko/react-native-teleport/issues/24 in terms
of what should be changed to make it working. But as a start point I'd
love to have an example that will show case the problem and will compare
it with proper layout in portals that are based on contexts and web.
This PR exactly adds it.

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- added example when teleported child has `flex: 1` style.

## 🤔 How Has This Been Tested?

Tested on iOS/Android/web.

## 📸 Screenshots (if appropriate):

<img width="1730" height="906" alt="image"
src="https://github.com/user-attachments/assets/bf26d7b6-12c0-494b-abb1-1b6f80e1ec3e"
/>

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed