docs: mention that lib is compatible only with new architecture (#33)
## 📜 Description

Mention in documentation explicitly that this package requires new
architecture.

## 💡 Motivation and Context

The development of this library started in September 2025. In Sep 2025
it was announced that RN sunsetting an old architecture and starting
from RN 0.82 it's impossible to disable the new architecture. The new
architecture was promoted to stable/beta from RN 0.77 which was released
in Jan 2025.

So when I started development of this package I was aiming to support
**only** new architecture. Technically it's possible to add a support
for old architecture because I don't use new architecture benefits yet,
but:
- it'll consume more resources/development hours;
- this library is not so popular/has some major bugs so for me it looks
more reasonable to stabilize it first. And most likely when library will
become popular by that time the old architecture will completely gone,
so adding support for old architecture may be waste of time

So taking everything from above into consideration - supporting old
architecture is not a priority at the moment, and keeping in mind that
old architecture is officially deprecated it may be not reasonable to
add support for it right now 🤷‍♂️

Closes https://github.com/kirillzyusko/react-native-teleport/issues/32

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### Docs

- mention in installation that lib is not compatible with old
architecture;

## 🤔 How Has This Been Tested?

Tested via preview.

## 📸 Screenshots (if appropriate):

<img width="989" height="451" alt="image"
src="https://github.com/user-attachments/assets/80a426cb-e329-438c-a4b1-3cddcb8ac404"
/>

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed