fix: cached size (#48)
## 📜 Description

Fixed an issue with intermittent incorrect layout size of `PortalHost`.

## 💡 Motivation and Context

The problem is that sometimes when we real layout metrics of portal host
we may get incorrect results.

I assume it stems from the fact that `adopt` may be gets called at any
point of time and there is no guarantee that layout phase has been
finished. So we may read layout that hasn't been completed yet, and as a
result we'll get incorrect values.

My initial idea was to move that logic with resizing into
`shadowTreeWillCommit` hook
https://github.com/kirillzyusko/react-native-teleport/pull/47 But even
there I still get invalid layout, though I assume, that
`shadowTreeWillCommit` gets called when layout has been fully
calculated.

Anyway, the solution with `shadowTreeWillCommit` hook requires more
changes and needs more investigation. So as an intermediate solution I
come up with this:

- we don't register direct nodes in the registry;
- instead we store only layout information (width/height);
- we update layout info from `layout` callback;
- as a result we'll consume only valid layout from the registry.

I think this approach may introduce a problem of "first frame lag". But
since we teleport almost all content to "root", it's safe to skip first
frames and get a cached layout. Again, solution is not perfect, but it
fixes current problem. Later on I'll revisit the solution 🤞

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### CI

- don't recreate AVD force;
- include `common` directory for e2e tests;

### C++

- `PortalRegistry` stores size instead of `ShadowNode` references;
- update `PortalRegistry` from `layout` method lifecycle.

## 🤔 How Has This Been Tested?

Tested via e2e tests and in client app.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed