feat: instagram reels (#51)
## 📜 Description

Added instagram feed -> reels transition (iOS style) that works on all
platforms: Android/iOS/web

## 💡 Motivation and Context

This is a WIP example. I wanted to check that re-parenting for video
works (and it works, but on Android we need to apply a patch to make it
working better). In this PR I installed many new deps, and because of
that I can not easily jump between branches, so I decided to merge this
PR as is and make it looking good later on.

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- linked `react-native-vector-icons`;
- linked `react-native-video`;
- added demo videos;
- added new screen for feed/reels;

## 🤔 How Has This Been Tested?

Tested on Pixel 7 Pro (Android 16)/Xiaomi Redmi Note 5 Pro Android 9) +
iPhone 16 Pro (iOS 26.0) + Chrome/Safari browsers.

## 📸 Screenshots (if appropriate):


https://github.com/user-attachments/assets/d07ff457-9864-4ba2-8ae5-64948339dc3d

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed